Move classes for .h to .cpp as it is now an implementation detail of higher methods
