+++ apps/openmw/mwstate/quicksavemanager.cpp
@@ -1,5 +1,31 @@
 #include "quicksavemanager.hpp"
 
+namespace MWState {
+    class NextSlotFinder {
+        std::string mSaveName;
+        unsigned int mMaxSaves;
+        unsigned int mSlotsVisited;
+        const Slot *mOldestSlotVisited;
+    private:
+        bool shouldCreateNewSlot() const;
+        bool isOldestSave(const Slot *compare) const;
+    public:
+        NextSlotFinder(const std::string &saveName, unsigned int maxSaves);
+        ///< A utility class to manage multiple quicksave slots
+        ///
+        /// \param saveName The name of the save ("QuickSave", "AutoSave", etc)
+        /// \param maxSaves The maximum number of save slots to create before recycling old ones
+
+        void visitSave(const Slot *saveSlot);
+        ///< Visits the given \a slot \a
+
+        const Slot *getNextQuickSaveSlot() const;
+        ///< Get the slot that the next quicksave should use.
+        ///
+        ///\return Either the oldest quicksave slot visited, or NULL if a new slot can be made
+    };
+}
+
 MWState::NextSlotFinder::NextSlotFinder(const std::string &saveName, unsigned int maxSaves)
   : mSaveName(saveName)
   , mMaxSaves(maxSaves)
@@ -37,6 +63,28 @@ const MWState::Slot *MWState::NextSlotFinder::getNextQuickSaveSlot() const
     return mOldestSlotVisited;
 }
 
+
+namespace MWState {
+    class LatestSlotFinder {
+        std::string mSaveName;
+        const Slot *mLatestSlotVisited;
+    private:
+        bool isLatestSave(const Slot *compare) const;
+    public:
+        LatestSlotFinder(const std::string &saveName);
+        ///< A utility class to manage multiple quicksave slots
+        ///
+        /// \param saveName The name of the save ("QuickSave", "AutoSave", etc)
+
+        void visitSave(const Slot *saveSlot);
+        ///< Visits the given \a slot \a
+
+        const Slot *getLatestSaveSlot() const;
+        ///< Get the latest quicksave.
+        ///
+        ///\return Either the latest quicksave slot visited, or NULL if there is no quicksave
+    };
+}
 
 MWState::LatestSlotFinder::LatestSlotFinder(const std::string &saveName)
     : mSaveName(saveName)
